% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_modparams.R
\name{extract_modparams}
\alias{extract_modparams}
\title{This function allows you to extract model parameters from the multilevel models}
\usage{
extract_modparams(result_list, nreps)
}
\arguments{
\item{result_list}{results from multilevel models}

\item{nreps}{number of replications}
}
\value{
the model parameters
}
\description{
This function allows you to extract model parameters from the multilevel models
}
