% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_para.R
\name{get_para}
\alias{get_para}
\title{This function allows you to simulate longitudinal ordinal data and estimate the parameters with multilevel models}
\usage{
get_para(
  n,
  numSample,
  numDays,
  assess,
  thresh,
  autoreg_coeff,
  crosslag_coeff,
  gamma_00,
  gamma_00_sd,
  gamma_01_sd
)
}
\arguments{
\item{n}{number of categories for the ordinal outcome variable}

\item{numSample}{number of participants}

\item{numDays}{number of days}

\item{assess}{four accessments per day}

\item{thresh}{thresholds for ordinal outcome}

\item{autoreg_coeff}{autoregressive coefficient}

\item{crosslag_coeff}{cross-lag coefficient}

\item{gamma_00}{fixed intercept}

\item{gamma_00_sd}{random intercept}

\item{gamma_01_sd}{random autoregressive cofficient}
}
\value{
the estimated cross-lag coefficient and its corresponding p-value
}
\description{
This function allows you to simulate longitudinal ordinal data and estimate the parameters with multilevel models
}
