% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param.R
\name{get_param}
\alias{get_param}
\title{This function allows you to simulate longitudinal ordinal data and estimate the parameters with multilevel models}
\usage{
get_param(
  n,
  numSample,
  numAssess,
  thresh_CON,
  autoreg_coeff,
  crosslag_coeff,
  crosslag_sk,
  gamma_00,
  gamma_00_sd,
  gamma_01_sd,
  gamma_02_sd,
  Compliance,
  crosslag_prior,
  ar_sk,
  corr
)
}
\arguments{
\item{n}{number of categories for the ordinal outcome variable}

\item{numSample}{number of participants}

\item{numAssess}{number of assessments}

\item{thresh_CON}{condition for thresholds: normal or skewed}

\item{autoreg_coeff}{autoregressive coefficient}

\item{crosslag_coeff}{cross-lag coefficient}

\item{crosslag_sk}{cross-lag skewness and kuertosis}

\item{gamma_00}{fixed intercept}

\item{gamma_00_sd}{random intercept}

\item{gamma_01_sd}{random autoregressive cofficient sd}

\item{gamma_02_sd}{random cross-lag coefficient sd}

\item{Compliance}{compliance rate in percentage}

\item{crosslag_prior}{prior for crosslag}

\item{ar_sk}{ar skewness and kuertosis}

\item{corr}{correlation between random ar and random crosslag}
}
\value{
the estimated cross-lag coefficient and its corresponding p-value
}
\description{
This function allows you to simulate longitudinal ordinal data and estimate the parameters with multilevel models
}
