% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param.R, R/get_param_assess_per_day.R
\name{get_param}
\alias{get_param}
\title{This function allows you to simulate longitudinal ordinal data and estimate the parameters with multilevel models}
\usage{
get_param(
  n,
  numSample,
  numDays,
  assess,
  thresh,
  autoreg_coeff,
  crosslag_coeff,
  gamma_00,
  gamma_00_sd,
  gamma_01_sd
)

get_param(
  n,
  numSample,
  numDays,
  assess,
  thresh,
  autoreg_coeff,
  crosslag_coeff,
  gamma_00,
  gamma_00_sd,
  gamma_01_sd
)
}
\arguments{
\item{n}{number of categories for the ordinal outcome variable}

\item{numSample}{number of participants}

\item{numDays}{number of days}

\item{assess}{four accessments per day}

\item{thresh}{thresholds for ordinal outcome}

\item{autoreg_coeff}{autoregressive coefficient}

\item{crosslag_coeff}{cross-lag coefficient}

\item{gamma_00}{fixed intercept}

\item{gamma_00_sd}{random intercept}

\item{gamma_01_sd}{random autoregressive cofficient}

\item{numAssess}{number of assessments}

\item{gamma_02_sd}{random cross-lag coefficient sd}

\item{Compliance}{compliance rate in percentage}
}
\value{
the estimated cross-lag coefficient and its corresponding p-value

the estimated cross-lag coefficient and its corresponding p-value
}
\description{
This function allows you to simulate longitudinal ordinal data and estimate the parameters with multilevel models

This function allows you to simulate longitudinal ordinal data and estimate the parameters with multilevel models
}
