% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_power.R
\name{ord_power}
\alias{ord_power}
\title{This function allows you to simulate longitudinal ordinal data and estimate the parameters with multilevel models for a nubmer of times}
\usage{
ord_power(
  n,
  numSample,
  numDays,
  assess,
  thresh,
  autoreg_coeff,
  crosslag_coeff,
  gamma_00,
  gamma_00_sd,
  gamma_01_sd,
  reps
)
}
\arguments{
\item{n}{number of categories for the ordinal outcome variable}

\item{numSample}{number of participants}

\item{numDays}{number of days}

\item{assess}{four accessments per day}

\item{thresh}{thresholds for ordinal outcome}

\item{autoreg_coeff}{autoregressive coefficient}

\item{crosslag_coeff}{cross-lag coefficient}

\item{gamma_00}{fixed intercept}

\item{gamma_00_sd}{random intercept}

\item{gamma_01_sd}{random autoregressive cofficient}

\item{reps}{number of replications}
}
\value{
the mean of estimated cross-lag coefficient and percentage of significant effects from number of replicated simulations
}
\description{
This function allows you to simulate longitudinal ordinal data and estimate the parameters with multilevel models for a nubmer of times
}
