% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_power.R
\name{ord_power}
\alias{ord_power}
\title{This function allows you to simulate longitudinal ordinal data and estimate the parameters with multilevel models for a nubmer of times}
\usage{
ord_power(
  n,
  numSample,
  numAssess,
  thresh_CON,
  autoreg_coeff,
  crosslag_coeff,
  crosslag_skew = 0,
  crosslag_kurt = 3,
  gamma_00,
  gamma_00_sd,
  gamma_01_sd,
  gamma_02_sd,
  Compliance = 1,
  ar_skew = 0,
  ar_kurt = 3,
  corr = 1,
  reps = 100
)
}
\arguments{
\item{n}{number of categories for the ordinal outcome variable}

\item{numSample}{number of participants}

\item{numAssess}{number of assessments}

\item{thresh_CON}{threshold conditions, 1 represents normal thresholds, 2 represents right-skewed thresholds}

\item{autoreg_coeff}{autoregressive coefficient}

\item{crosslag_coeff}{cross-lag coefficient}

\item{crosslag_skew}{cross-lag skewness}

\item{crosslag_kurt}{cross-lag kurtosis}

\item{gamma_00}{fixed intercept}

\item{gamma_00_sd}{random intercept}

\item{gamma_01_sd}{random autoregressive cofficient sd}

\item{gamma_02_sd}{random cross-lag coefficient sd}

\item{Compliance}{compliance rate in percentage}

\item{ar_skew}{ar skewness}

\item{ar_kurt}{ar kurtosis}

\item{corr}{correlation between random ar and random crosslag, 1 indicating correlation of 0.5, 2 indicating no corrrelation}

\item{reps}{number of replications}
}
\value{
the mean of estimated cross-lag coefficient, the percentage of significant effects, the percentage of non-significant effects, total replication number, percentage of replications that produced errors, and percentage of replications that produced warnings
}
\description{
This function allows you to simulate longitudinal ordinal data and estimate the parameters with multilevel models for a nubmer of times
}
\examples{
ord_power(n = 5,
numSample = 35,
numAssess = 73,
thresh_CON = 2,
autoreg_coeff = 0.919,
crosslag_coeff = 0.4,
gamma_00 = 0,
gamma_00_sd = 1.043,
gamma_01_sd = 0.591,
gamma_02_sd = 0.13,
Compliance = 1,
reps=10)
}
